3/8

+++ admin/products/products_list.php
@@ -2,7 +2,7 @@
 require_once "../dao/products.php";
 require_once "../dao/categories.php";
 
-// hiện session
+// hiện cookie
 if (isset($_COOKIE['add-product'])) {
     echo '<div class="alert alert-success" role="alert">
             ' . $_COOKIE['add-product'] . '
@@ -68,10 +68,9 @@ $products = select_products_by_page($start, $limit);
         </tr>
     </thead>
     <tbody>
-        <?php foreach ($products as $product): ?>
+        <?php foreach ($products as $product) : ?>
             <tr>
-                <th scope="col"><input class="form-check-input" type="checkbox" value="<?= $product['product_id'] ?>"
-                        name="check_hh"></th>
+                <th scope="col"><input class="form-check-input" type="checkbox" value="<?= $product['product_id'] ?>" name="check_hh[]"></th>
                 <th scope="col">
                     <?= $product['product_id'] ?>
                 </th>
@@ -96,11 +95,8 @@ $products = select_products_by_page($start, $limit);
                     <?= $product['sold'] ?>
                 </th>
                 <th scope="col">
-                    <a href="index.php?act=edit_product&product_id=<?= $product['product_id'] ?>"><button
-                            class="button">Sửa</button></a>
-                    <a onclick="return confirm('Bạn có muốn xóa hàng này không')"
-                        href="products/delete_product.php?product_id=<?= $product['product_id'] ?>"><button
-                            class="button">Xóa</button></a>
+                    <a href="index.php?act=edit_product&product_id=<?= $product['product_id'] ?>"><button class="btn btn-outline-primary">Sửa</button></a>
+                    <a onclick="return confirm('Bạn có muốn xóa hàng này không')" href="products/delete_product.php?product_id=<?= $product['product_id'] ?>"><button class="btn btn-outline-primary">Xóa</button></a>
                 </th>
             </tr>
         <?php endforeach ?>
@@ -125,9 +121,6 @@ $products = select_products_by_page($start, $limit);
 
 <section class="mt-1">
     <div class="button">
-        <button type="submit" id="check-all" class="btn btn-outline-primary">Chọn tất cả</button>
-        <button type="button" id="uncheck" class="btn btn-outline-primary">Bỏ chọn tất cả</button>
-        <button type="button" class="btn btn-outline-primary">Xóa các mục chọn</button>
         <a href="index.php?act=add-product"><button type="button" class="btn btn-outline-primary">Nhập thêm</button></a>
     </div>
 </section>
